+++ src/main/java/cc/polyfrost/oneconfig/gui/elements/config/ConfigCheckbox.java
@@ -95,7 +95,7 @@ public class ConfigCheckbox extends BasicOption {
     }
 
     @Override
-    protected int getNameX(int x) {
+    protected float getNameX(int x) {
         return x + 32;
     }
